examples: improve code after seminar №01
Use double instead of float as number of rubles.
It turned out that 32 bit float is not enough
for 8 decimal digits. According to IEEE754 a 32-bit float
has only 23 binary digits.

Required number of bits for 8 decimal digits:
log_2(10)*8 = 26.5 > 23

Also add link to VSCode tutorial.

+++ examples/01_dollar/dollar.c
@@ -4,7 +4,7 @@ const unsigned DOLLAR_START  = 0;
 const unsigned DOLLAR_FINISH = 1000;
 const unsigned DOLLAR_STEP   = 100;
 
-const float RUBLES_IN_DOLLAR = 97.8439;
+const double RUBLES_IN_DOLLAR = 97.8439;
 
 int main(void)
 {
@@ -12,9 +12,9 @@ int main(void)
         num_dollars <= DOLLAR_FINISH;
         num_dollars += DOLLAR_STEP)
     {
-        float num_rubles = RUBLES_IN_DOLLAR * num_dollars;
+        double num_rubles = RUBLES_IN_DOLLAR * num_dollars;
 
-        printf("%4d USD = %9.03f RUB\n", num_dollars, num_rubles);
+        printf("%4d USD = %10.3f RUB\n", num_dollars, num_rubles);
     }
 
     return 0;
